requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.1
python-multipart==0.0.17
cryptography==44.0.0
pydantic-settings==2.6.1
email-validator==2.2.0
google-auth==2.36.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
google-api-python-client==2.154.0
selenium==4.27.1
python-dotenv==1.0.1
icalendar==6.0.1
requests==2.32.3
